# Conflicts:
#	src/controller/BookCRUDWindowController.java
#	src/controller/BookOptionWindowController.java
#	src/controller/DeleteAccountAdminController.java
#	src/controller/DeleteComentWindowController.java
#	src/main/Main.java
#	src/model/ClassDAO.java
#	src/model/DBImplementation.java
#	src/view/BookCRUDWindow.fxml

+++ src/model/CommentateId.java
@@ -31,7 +31,7 @@ public class CommentateId implements Serializable {
         return isbnBook;
     }
 
-    public void setIsbnBook(int isbnBook) {
+    public void setIsbnBook(long isbnBook) {
         this.isbnBook = isbnBook;
     }
